[NOT READY | NO LINT ERRORS] renamed, reorg, add loaders
