+++ apps/web/src/components/home/stack/Overlay.tsx
