+++ apps/web/src/components/home/stack/SlideShow.tsx
