+++ apps/web/src/components/home/stack/TenantStack.tsx
