+++ apps/web/src/components/posts/PostPage.tsx
@@ -1,4 +1,4 @@
-import { DisplayPost } from "../users/DisplayProperty";
+import { DisplayPost } from "../users/posts/DisplayProperty";
 import { trpc } from "../../utils/trpc";
 import { useEffect } from "react";
