+++ apps/web/src/components/shared/Loader.tsx
@@ -16,7 +16,7 @@ export const Loader = () => {
           fill="currentFill"
         />
       </svg>
-      <span className="sr-only">Chargement...</span>
+      <span className="sr-only">Loading...</span>
     </div>
   );
 };
