+++ apps/web/src/components/users/matches/MatchesPage.tsx
@@ -1,8 +1,9 @@
 import { Role } from "@prisma/client";
 import { Header } from "../Header";
 import { TenantList } from "../TenantList";
-import { PostList } from "../PostList";
+import { PostList } from "../posts/PostList";
 import { trpc } from "../../../utils/trpc";
+import { Loader } from "../../shared/Loader";
 
 export interface MatchesPageProps {
   userId: string;
@@ -12,7 +13,7 @@ export function MatchesPage({ userId }: MatchesPageProps) {
   const { data: session, isLoading } = trpc.auth.getSession.useQuery();
 
   if (isLoading) {
-    return <div>Loading...</div>;
+    return <Loader />;
   }
 
   if (!session) {
