+++ apps/web/src/components/users/posts/DeletePostButton.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { trpc } from "../../utils/trpc";
+import { trpc } from "../../../utils/trpc";
 import { useRouter } from "next/router";
 import { useSession } from "@clerk/nextjs";
