+++ apps/web/src/pages/support.tsx
@@ -1,28 +1,11 @@
 import { Role } from "@prisma/client";
-import { useRouter } from "next/router";
 import { LoggedLayout } from "../components/shared/layout/LoggedLayout";
-import { trpc } from "../utils/trpc";
 import { Support } from "../components/moderation/Support";
-import { Loader } from "../components/moderation/Loader";
 
 export default function SupportPage() {
   return (
-    <LoggedLayout title="Support | Leace">
-      <AdminCheck>
-        <Support />
-      </AdminCheck>
+    <LoggedLayout title="Support | Leace" roles={[Role.ADMIN]}>
+      <Support />
     </LoggedLayout>
   );
 }
-
-const AdminCheck = ({ children }: { children: React.ReactNode }) => {
-  const { data: session } = trpc.auth.getSession.useQuery();
-  const router = useRouter();
-
-  if (!session || session.role != Role.ADMIN) {
-    router.push("/");
-    return <Loader />;
-  }
-
-  return <>{children}</>;
-};
